Move OpenPgpAppSelectDialog to new account settings UI package
