+++ k9mail/src/main/java/com/fsck/k9/ui/settings/account/OpenPgpAppSelectDialog.java
@@ -1,4 +1,4 @@
-package com.fsck.k9.activity.setup;
+package com.fsck.k9.ui.settings.account;
 
 
 import java.util.ArrayList;
